add getDistanceToNode function

+++ src/main/java/Tile.java
@@ -33,7 +33,7 @@ class Tile implements Comparable<Tile> {
 
     // to get the distance between two tiles we need to convert to the cube system and then do the calculation
     // see: https://www.redblobgames.com/grids/hexagons/#distances
-    double getDistance(Tile otherTile) {
+    int getDistance(Tile otherTile) {
 
 
         Cube a = new Cube(this);
